feat!: BREAKING move all code to `ts` category
This will solve some issues with name collision in my projects using utils. Normally I just create a ts directory so this will function the same way
