+++ src/ts/array-sum.ts
