+++ src/ts/array-to-map.ts
