+++ src/ts/is-number.test.ts
