+++ src/ts/is-number.ts
