+++ src/ts/math/gcf.test.ts
