+++ src/ts/math/gcf.ts
