+++ src/ts/math/index.ts
