+++ src/ts/math/triangles.ts
