+++ src/ts/rand.test.ts
