+++ src/ts/rand.ts
