+++ src/ts/result.test.ts
