+++ src/ts/result.ts
