+++ src/ts/stopwatch.ts
