+++ src/ts/strip-ansi.test.ts
