+++ src/ts/strip-ansi.ts
