+++ src/ts/truncate.test.ts
