+++ src/ts/truncate.ts
